removing capacitor base class and integrating into phase_capacitor
